feat(sidebar): implement AppSidebar component to enhance navigation and user experience refactor(sidebar): rename 'Projects' to 'Workspaces' and update related labels for consistency chore(sidebar): remove unused sidebar components to streamline codebase and improve maintainability
