--- client/src/components/sidebar/AppSideBarHeader.tsx
@@ -1,10 +1,0 @@
-import {SidebarHeader} from '@/components/shadcn/ui/sidebar';
-import {FC} from 'react';
-
-export const AppSideBarHeader: FC = () => {
-    return (
-        <SidebarHeader className="text-xl font-bold border-b">
-            AIGraphr
-        </SidebarHeader>
-    );
-};
