--- client/src/components/sidebar/sidebar.types.ts
@@ -1,15 +1,0 @@
-import {IconType} from 'react-icons';
-
-export interface AppSideBarItem {
-    icons: {
-        active: IconType;
-        inactive: IconType;
-    };
-
-    links: {
-        activate: string;
-        isActive: string;
-    };
-
-    title: string;
-}
